:heavy_minus_sign: chore: remove @c2d2c/utils
