+++ package.json
@@ -62,7 +62,6 @@
     "@antv/dw-random": "^1",
     "@babel/runtime": "^7.21.5",
     "@c2d2c/types": "^1",
-    "@c2d2c/utils": "^1",
     "@dnd-kit/core": "^6",
     "@dnd-kit/modifiers": "^6",
     "@dnd-kit/sortable": "^7",
