+++ src/ComponentAsset/ComponentAsset.tsx
@@ -1,8 +1,8 @@
 /*eslint no-invalid-this: "error"*/
-import { getDefaultValueFromSchema } from '@c2d2c/utils';
 import { FC, ReactNode } from 'react';
 import { StoreApi } from 'zustand';
 import type { UseBoundStore } from 'zustand/react';
+import { getDefaultValueFromSchema } from '../utils';
 
 import type { EditorMode } from '../ProEditor';
 import type { AssetModels, CanvasRule, CodeEmitter, ComponentAssetParams } from './types';
