+++ src/ComponentAsset/types/index.ts
@@ -1,5 +1,5 @@
+export * from '../../types/schema';
 export * from './asset';
 export * from './code';
 export * from './model';
 export * from './render';
-export * from './schema';
