+++ src/utils/index.ts
@@ -1,1 +1,2 @@
 export * from './autoId';
+export * from './schema';
